refactor: Declare function in abstract class
