+++ src/main/java/command/parser/InsuranceParser.java
@@ -1,10 +1,5 @@
 package command.parser;
 
-import domain.enums.EmployeeType;
-import domain.enums.InsuranceType;
-import domain.enums.Treaty;
-import java.util.Arrays;
-
 public class InsuranceParser extends Parser {
 
 
@@ -24,26 +19,6 @@ public class InsuranceParser extends Parser {
   @Override
   public void initialize() {
     System.out.println("Insurance");
-  }
-
-  public InsuranceType getInsuranceType() {
-    System.out.println("보험종류를 선택해주세요");
-    Arrays.stream(InsuranceType.values()).forEach(type -> {
-      System.out.println(
-          Integer.toString(type.ordinal() + 1)
-              .concat(". ")
-              .concat(type.toString()));
-    });
-
-    System.out.print("보험종류 : ");
-    while (!sc.hasNextInt()) {
-      sc.next();
-      System.out.println("Please enter a number within the range.");
-    }
-
-    int insuranceType = sc.nextInt();
-    return (insuranceType < 1 || insuranceType > EmployeeType.values().length) ? getInsuranceType()
-        : InsuranceType.values()[insuranceType - 1];
   }
